add unit tests for eod/[date]
